Increase file loading speed
File loading is now 26 times faster on the test file. New strategy is
to load entirely the file instead of reading line by line. Test the
definition type on this and then only read line by line to create the
array of constraints. Also the text of the file is directly created
from the loaded file instead of incrementally while reading line by
line. (Thanks to qcachegrind, cprofile  and python conversion script
between the two !).