Separation of loading and parsing
Parsing of constraints is now done by a specific class. There is a
class for each type of constraint inherited from a generic parsing
class. In those, I used generators, class attributes and static
methods, this is a much more pythonic way to do. CYANA types should be
further tested.