Separated window for NOE Data view
As described in previous commit, NOE Data can now be consulted in a
separated window. All necessary code has been transferred from NMRCore
/ NOEDrawing to new classes NOEDataController / NOEDataViewer.
Bonus : Corrected a bug where when the NOE where restricted to a
certain subset, this was not reflected in the reflection, created a new
method intersection in ConstraintSetManager for that purpose.