New datalist controllers, AtomSet class removed
In order to allow segued to be displayed several changes have been
done. AtomSet has been replaced by an namedTuple (this has incidence
all along the chain of processing from loading to viewing). Two new
classes acting as datalist controllers have been created in order to
ease the transmission of data between the constraintManager and the
data viewer. There might be some more work to do, but at least it is
functional. It should be tested with a structure containing several
segments / chains.