Major bug in density calculation and display solved
Constraint density was calculated at atom level while residue level was required. Therefore the displayed density was the one of the latest atom  in the list for a residue.
As well, simple atom selection in structure was not handling properly segid mismatch between pymol and their definition in the constraint file. Additional method has been added in atomList.