Selection GUI deep refactoring
The code of the NOE Selection GUI has been refactored in 3 files, one object per file. The refactoring has been done with MVC pattern in mind, similar to TableView of Cocoa Framework. This was made possible by the delegates decorators by Tom Hennigan.
Overall the user experience is not changed, but it should be simpler to maintain and enhance.